+++ packages/ts/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Add document-wide `objectAdditionalProperties` output policy for `object()`
+  schemas.
+
 ## 1.1.0 - 2026-05-27
 
 ### Added